integration_2023_w22

See merge request oai/openairinterface5g!2145

* !2036 CPPCHECK fixes for printf issues
* !2110 Improvements in CSI-RS functions
* !2132 NR fix problems found with undefined behavior sanitizer
* !2136 CI: Fix of the previous modification in iperf module (MR2100)
* !2137 chore(ci): removing OAI-Rel14-CUPS old bare-metal 4G core installation variant
* !2141 remove warning for non existing RNTI
* !2144 Build the CUUP on the cluster and minor Dockerfile fixes
* !2122 reworking MAC UE Search Space and Coreset configuration v2
* !1828 asn1c: use better version of asn1c
* !2130 ci: activate DRB ciphering and integrity protection
* !2140 F1: do not assert on unknown F1 message
* !2134 Cleanup RRC SRB data structures and make 2 PDU sessions functional

+++ openair1/PHY/LTE_TRANSPORT/dci_tools_common.c
@@ -933,7 +933,7 @@ uint8_t subframe2harq_pid(LTE_DL_FRAME_PARMS *frame_parms,uint32_t frame,uint8_t
     }
   }
 
-  AssertFatal(ret != 255, "invalid harq_pid(%d) at SFN/SF = %d/%d\n", (int8_t)ret, frame, subframe);
+  AssertFatal(ret != 255, "invalid harq_pid(%u) at SFN/SF = %u/%u\n", ret, frame, subframe);
   return ret;
 }
